+++ xtea4/counter.py
@@ -40,7 +40,7 @@ class Counter:
     """Small counter for CTR mode, based on arrays
     Example:
     
-        >>> from xtea.counter import Counter
+        >>> from xtea4.counter import Counter
         >>> nonce = b"$2dUI84e" # This should be random
         >>> c = Counter(nonce)
         >>> c()